adding

+++ codeforces/682/a.cpp
@@ -6,10 +6,10 @@ using namespace std;
 int main(int argc, char const *argv[]) {
 	long long n,m;
 	cin>>n>>m;
-	int a=max(n,m);
-	int b=n+m-a;
-	int x=0;
-	for(int i=1;i<=b;i++){
+	long long int a=max(n,m);
+	long long int b=n+m-a;
+	long long int x=0;
+	for(long long int i=1;i<=b;i++){
 		x+=((a+i)/5)-(i/5);
 	}
 	cout<<x;
